Cast SimulationEngine.getTime() calls to int so that they can be used to look up the simulation year in the input .xls files.

+++ demo07/src/it/unito/jas2/demo07/model/Person.java
@@ -303,7 +303,7 @@ public class Person implements Comparable<Person>, EventListener, IDoubleSource,
 		
 		MultiKeyCoefficientMap map = ( gender.equals(Gender.Male) ? Parameters.getpDeathM() : Parameters.getpDeathF() ) ;
 		
-		double deathProbability = ((Number) map.getValue(this.age, SimulationEngine.getInstance().getTime())).doubleValue();
+		double deathProbability = ((Number) map.getValue(this.age, (int)SimulationEngine.getInstance().getTime())).doubleValue();
 		
 		if ( RegressionUtils.event(deathProbability) ) {
 			// update partner's status
@@ -340,7 +340,7 @@ public class Person implements Comparable<Person>, EventListener, IDoubleSource,
 		if ((this.gender.equals(Gender.Female)) && (this.age >= 15) && (this.age <= 50)) {
 			double birthProbability;
 			try{
-				birthProbability = ((Number) Parameters.getpBirth().getValue(this.age, SimulationEngine.getInstance().getTime())).doubleValue();
+				birthProbability = ((Number) Parameters.getpBirth().getValue(this.age, (int)SimulationEngine.getInstance().getTime())).doubleValue();
 				if ( RegressionUtils.event(birthProbability) ) {
 
 					@SuppressWarnings("unused")
